requirements.txt
requests
python-dotenv
pandas
pydantic